Changed types to size_t on xbd_multipacket.c

+++ xbd_multipacket.h
@@ -22,11 +22,14 @@ const void *tgt_data;
 
 size_t XBD_genInitialMultiPacket(struct xbd_multipkt_state *state, const void *srcdata, size_t srclen, void *dstbuf, const char *cmd_code, uint32_t addr, uint32_t type);
 
-size_t XBD_genSucessiveMultiPacket(struct xbd_multipkt_state *state, void *dstbuf, uint32_t dstlenmax, const char *cmd_code);
+size_t XBD_genSucessiveMultiPacket(struct xbd_multipkt_state *state, void
+        *dstbuf, size_t dstlenmax, const char *cmd_code);
 
-int XBD_recInitialMultiPacket(struct xbd_multipkt_state *state, const void *recdata, uint32_t reclen, const char *cmd_code, bool hastype, bool hasaddr);
+int XBD_recInitialMultiPacket(struct xbd_multipkt_state *state, const void
+        *recdata, size_t reclen, const char *cmd_code, bool hastype, bool hasaddr);
 
-int XBD_recSucessiveMultiPacket(struct xbd_multipkt_state *state, const void *recdata, uint32_t reclen, void *dstbuf, uint32_t dstlenmax, const char *code);
+int XBD_recSucessiveMultiPacket(struct xbd_multipkt_state *state, const void
+        *recdata, size_t reclen, void *dstbuf, size_t dstlenmax, const char *code);
 
 
 #endif /* _XBD_MULTIPACKET_H */
